Move covariance_term.h and covariance_functions.h to covariance_function.h
