+++ albatross/covariance_functions/covariance_function.h
